move to shared: keyword-context/index
