+++ client/cody/src/keyword-context/local-keyword-context-fetcher.ts
@@ -5,7 +5,7 @@ import { removeStopwords } from 'stopword'
 import StreamValues from 'stream-json/streamers/StreamValues'
 import * as vscode from 'vscode'
 
-import { KeywordContextFetcher, KeywordContextFetcherResult } from '.'
+import { KeywordContextFetcher, KeywordContextFetcherResult } from '@sourcegraph/cody-shared/src/keyword-context'
 
 const fileExtRipgrepParams = ['-Tmarkdown', '-Tyaml', '-Tjson', '-g', '!*.lock']
