+++ client/cody/src/test/unit/mocks.ts
@@ -1,9 +1,9 @@
 import { EmbeddingsSearch } from '@sourcegraph/cody-shared/src/embeddings'
 import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
+import { KeywordContextFetcher, KeywordContextFetcherResult } from '@sourcegraph/cody-shared/src/keyword-context'
 import { EmbeddingsSearchResults } from '@sourcegraph/cody-shared/src/sourcegraph-api/graphql'
 
 import { ActiveTextEditor, ActiveTextEditorSelection, ActiveTextEditorVisibleContent, Editor } from '../../editor'
-import { KeywordContextFetcher, KeywordContextFetcherResult } from '../../keyword-context'
 
 export class MockEmbeddingsClient implements EmbeddingsSearch {
     constructor(private mocks: Partial<EmbeddingsSearch> = {}) {}
